employee-app/scripts/requirements.txt
gspread==6.1.2
oauth2client==4.1.3
pymongo==4.7.2
python-dotenv==1.0.1
